Force a test on a specific worker
Typical Buildbot behavior is to have all workers run that are associated with a
build.  In order to start a test only on a specific worker, a ForceScheduler
needs to be created for every test that a worker is allowed to run.  Then a
corresponding BuilderConfig is needed for each worker and test.

Also need to play extract just the test name from the builder when determine
which custom GUC file to use.

While updating force-tests-rpc.py to also work with the new naming scheme, also
add a few upgrades:

* use argparse
* add a flag to limit the number of tests queued
* add flag to specify tests to run
* add flag to specify workers to queue tests against